+++ packages/tpetra/core/test/ImportExport2/ImportExport2_UnitTests.cpp
@@ -871,437 +871,6 @@ namespace {
   }
 
 
-  template <typename Scalar, typename LO, typename GO>
-  class MultiVectorTransferFixture {
-  private:
-    using map_type = Map<LO, GO>;
-    using mv_type = MultiVector<Scalar, LO, GO>;
-
-  public:
-    MultiVectorTransferFixture(Teuchos::FancyOStream& o, bool& s)
-      : out(o),
-        success(s),
-        comm(getDefaultComm()),
-        numProcs(comm->getSize()),
-        myRank(comm->getRank())
-    { }
-
-    ~MultiVectorTransferFixture() { }
-
-    bool shouldSkipTest() {
-      return isSerial();
-    }
-
-    void setupMultiVectors(int collectRank) {
-      const GO indexBase = 0;
-      const Tpetra::global_size_t INVALID = Teuchos::OrdinalTraits<Tpetra::global_size_t>::invalid();
-
-      const size_t sourceNumLocalElements = 3;
-      sourceMap = rcp(new map_type(INVALID, sourceNumLocalElements, indexBase, comm));
-      sourceMV = rcp(new mv_type(sourceMap, 1));
-      sourceMV->randomize();
-
-      const size_t totalElements = numProcs*sourceNumLocalElements;
-      const size_t targetNumLocalElements = (myRank == collectRank) ? totalElements : 0;
-      targetMap = rcp(new map_type(INVALID, targetNumLocalElements, indexBase, comm));
-      targetMV = rcp(new mv_type(targetMap, 1));
-      targetMV->putScalar(Teuchos::ScalarTraits<Scalar>::zero());
-    }
-
-    template <typename TransferMethod>
-    void performTransfer(const TransferMethod& transfer) {
-      transfer(sourceMV, targetMV);
-    }
-
-    template <typename ReferenceSolution>
-    void checkResults(const ReferenceSolution& referenceSolution) {
-      RCP<const mv_type> referenceMV = referenceSolution.generateWithClassicalCodePath(sourceMV, targetMap);
-      compareMultiVectors(targetMV, referenceMV);
-    }
-
-  private:
-    void compareMultiVectors(RCP<const mv_type> resultMV, RCP<const mv_type> referenceMV) {
-      auto data = resultMV->getLocalViewHost(Tpetra::Access::ReadOnly);
-      auto referenceData = referenceMV->getLocalViewHost(Tpetra::Access::ReadOnly);
-
-      for (GO globalRow = targetMap->getMinGlobalIndex(); globalRow <= targetMap->getMaxGlobalIndex(); ++globalRow) {
-        const LO localRow = targetMap->getLocalElement(globalRow);
-        TEST_EQUALITY(data(localRow, 0), referenceData(localRow, 0));
-      }
-    }
-
-    bool isSerial() {
-      if (numProcs < 2) {
-        out << "This test is only meaningful if running with multiple MPI "
-               "processes, but you ran it with only 1 process." << endl;
-        return true;
-      }
-      else return false;
-    }
-
-    Teuchos::FancyOStream& out;
-    bool& success;
-
-    RCP<const Comm<int>> comm;
-    const int numProcs;
-    const int myRank;
-
-    RCP<const map_type> sourceMap;
-    RCP<const map_type> targetMap;
-
-    RCP<mv_type> sourceMV;
-    RCP<mv_type> targetMV;
-  };
-
-  template <typename Scalar, typename LO, typename GO>
-  class ReferenceImport {
-  private:
-    using map_type = Map<LO, GO>;
-    using mv_type = MultiVector<Scalar, LO, GO>;
-
-  public:
-    RCP<mv_type> generateWithClassicalCodePath(RCP<mv_type> sourceMV, RCP<const map_type> targetMap) const {
-      RCP<const map_type> sourceMap = sourceMV->getMap();
-      Import<LO, GO> importer(sourceMap, targetMap, getImportParameterList());
-      expertSetRemoteLIDsContiguous(importer, false);
-      TEUCHOS_ASSERT(!importer.areRemoteLIDsContiguous());
-
-      RCP<mv_type> referenceMV = rcp(new mv_type(targetMap, 1));
-      referenceMV->putScalar(Teuchos::ScalarTraits<Scalar>::zero());
-      referenceMV->doImport(*sourceMV, importer, INSERT);
-      TEUCHOS_ASSERT(!referenceMV->importsAreAliased());
-
-      return referenceMV;
-    }
-  };
-
-  template <typename Scalar, typename LO, typename GO>
-  class ForwardImport {
-  private:
-    using mv_type = MultiVector<Scalar, LO, GO>;
-
-  public:
-    void operator()(RCP<mv_type> sourceMV, RCP<mv_type> targetMV) const {
-      Import<LO, GO> importer(sourceMV->getMap(), targetMV->getMap(), getImportParameterList());
-      targetMV->doImport(*sourceMV, importer, INSERT);
-    }
-  };
-
-  TEUCHOS_UNIT_TEST_TEMPLATE_3_DECL( MultiVectorTransfer, asyncImport, LO, GO, Scalar )
-  {
-    MultiVectorTransferFixture<Scalar, LO, GO> fixture(out, success);
-    if (fixture.shouldSkipTest()) return;
-
-    fixture.setupMultiVectors(0);
-    fixture.performTransfer(ForwardImport<Scalar, LO, GO>());
-    fixture.checkResults(ReferenceImport<Scalar, LO, GO>());
-  }
-
-
-  template <typename Scalar, typename LO, typename GO>
-  class CrsMatrixDiagonalTransferFixture {
-  private:
-    using map_type = Map<LO, GO>;
-    using crs_type = CrsMatrix<Scalar, LO, GO>;
-
-  public:
-    CrsMatrixDiagonalTransferFixture(Teuchos::FancyOStream& o, bool& s)
-      : out(o),
-        success(s),
-        comm(getDefaultComm()),
-        numProcs(comm->getSize()),
-        myRank(comm->getRank())
-    { }
-
-    ~CrsMatrixDiagonalTransferFixture() { }
-
-    bool shouldSkipTest() {
-      return isSerial();
-    }
-
-    void setupMatrices(int collectRank) {
-      const GO indexBase = 0;
-      const Tpetra::global_size_t INVALID = Teuchos::OrdinalTraits<Tpetra::global_size_t>::invalid();
-
-      const size_t targetNumLocalElements = 3;
-      const size_t totalElements = numProcs*targetNumLocalElements;
-      const size_t sourceNumLocalElements = (myRank == collectRank) ? totalElements : 0;
-
-      sourceMap = rcp(new map_type(INVALID, sourceNumLocalElements, indexBase, comm));
-      sourceMat = rcp(new crs_type(sourceMap, 1, StaticProfile, getCrsMatrixParameterList()));
-
-      if (sourceNumLocalElements != 0) {
-        for (GO row = sourceMap->getMinGlobalIndex(); row <= sourceMap->getMaxGlobalIndex(); row++) {
-          sourceMat->insertGlobalValues(row, tuple<GO>(row), tuple<Scalar>(row));
-        }
-      }
-      sourceMat->fillComplete();
-
-      targetMap = rcp(new map_type(INVALID, targetNumLocalElements, indexBase, comm));
-      targetMat = rcp(new crs_type(targetMap, 1, StaticProfile, getCrsMatrixParameterList()));
-    }
-
-    template <typename TransferMethod>
-    void performTransfer(const TransferMethod& transfer) {
-      transfer(sourceMat, targetMat);
-    }
-
-    template <typename ReferenceSolution>
-    void checkResults(const ReferenceSolution& referenceSolution) {
-      RCP<const crs_type> referenceMat = referenceSolution.generateUsingAllInOneImport(sourceMat, targetMap);
-      checkMatrixIsDiagonal(targetMat);
-      checkMatrixIsDiagonal(referenceMat);
-      compareMatrices(targetMat, referenceMat);
-    }
-
-  private:
-    void checkMatrixIsDiagonal(RCP<const crs_type> matrix) {
-      for (GO gblRow = targetMap->getMinGlobalIndex ();
-          gblRow <= targetMap->getMaxGlobalIndex ();
-          ++gblRow) {
-        const LO lclRow = targetMap->getLocalElement (gblRow);
-
-        typename crs_type::local_inds_host_view_type lclInds;
-        typename crs_type::values_host_view_type lclVals;
-        matrix->getLocalRowView (lclRow, lclInds, lclVals);
-        TEST_EQUALITY_CONST(lclInds.size(), 1);
-        TEST_EQUALITY_CONST(lclVals.size(), 1);
-
-        if (lclInds.size () != 0) { // don't segfault in error case
-          TEST_EQUALITY(matrix->getColMap ()->getGlobalElement (lclInds[0]), gblRow);
-        }
-        if (lclVals.size () != 0) { // don't segfault in error case
-          TEST_EQUALITY(lclVals[0], as<Scalar> (gblRow));
-        }
-      }
-    }
-
-    void compareMatrices(RCP<const crs_type> resultMat, RCP<const crs_type> referenceMat) {
-      using size_type = typename Array<Scalar>::size_type;
-      using magnitude_type = typename ScalarTraits<Scalar>::magnitudeType;
-
-      using lids_type = typename crs_type::nonconst_local_inds_host_view_type;
-      using vals_type = typename crs_type::nonconst_values_host_view_type;
-
-      const magnitude_type tol = as<magnitude_type>(10)*ScalarTraits<magnitude_type>::eps();
-
-      lids_type resultRowIndices;
-      vals_type resultRowValues;
-      lids_type referenceRowIndices;
-      vals_type referenceRowValues;
-      for (LO localRow = targetMap->getMinLocalIndex(); localRow <= targetMap->getMaxLocalIndex(); ++localRow)
-      {
-        size_t resultNumEntries = resultMat->getNumEntriesInLocalRow(localRow);
-        size_t referenceNumEntries = referenceMat->getNumEntriesInLocalRow(localRow);
-        TEST_EQUALITY(resultNumEntries, referenceNumEntries);
-
-        if (resultNumEntries > as<size_t>(resultRowIndices.size())) {
-          Kokkos::resize(resultRowIndices, resultNumEntries);
-          Kokkos::resize(resultRowValues, resultNumEntries);
-        }
-        if (referenceNumEntries > as<size_t>(referenceRowIndices.size())) {
-          Kokkos::resize(referenceRowIndices, referenceNumEntries);
-          Kokkos::resize(referenceRowValues, referenceNumEntries);
-        }
-
-        resultMat->getLocalRowCopy(localRow, resultRowIndices, resultRowValues, resultNumEntries);
-        referenceMat->getLocalRowCopy(localRow, referenceRowIndices, referenceRowValues, referenceNumEntries);
-
-        Tpetra::sort2(resultRowIndices, resultRowIndices.extent(0), resultRowValues);
-        Tpetra::sort2(referenceRowIndices, referenceRowIndices.extent(0), referenceRowValues);
-
-        for (size_type k = 0; k < static_cast<size_type>(resultNumEntries); ++k) {
-          TEST_EQUALITY(resultRowIndices[k], referenceRowIndices[k]);
-          TEST_FLOATING_EQUALITY(resultRowValues[k], referenceRowValues[k], tol);
-        }
-      }
-    }
-
-    bool isSerial() {
-      if (numProcs < 2) {
-        out << "This test is only meaningful if running with multiple MPI "
-               "processes, but you ran it with only 1 process." << endl;
-        return true;
-      }
-      else return false;
-    }
-
-    Teuchos::FancyOStream& out;
-    bool& success;
-
-    RCP<const Comm<int>> comm;
-    const int numProcs;
-    const int myRank;
-
-    RCP<const map_type> sourceMap;
-    RCP<const map_type> targetMap;
-
-    RCP<crs_type> sourceMat;
-    RCP<crs_type> targetMat;
-  };
-
-  template <typename Scalar, typename LO, typename GO>
-  class ReferenceImportMatrix {
-  private:
-    using map_type = Map<LO, GO>;
-    using crs_type = CrsMatrix<Scalar, LO, GO>;
-
-  public:
-    RCP<crs_type> generateUsingAllInOneImport(RCP<crs_type> sourceMat, RCP<const map_type> targetMap) const {
-      RCP<const map_type> sourceMap = sourceMat->getMap();
-      Import<LO, GO> importer(sourceMap, targetMap, getImportParameterList());
-
-      Teuchos::ParameterList dummy;
-      RCP<crs_type> referenceMat = Tpetra::importAndFillCompleteCrsMatrix<crs_type>(
-                                      sourceMat, importer, Teuchos::null, Teuchos::null, rcp(&dummy,false));
-      return referenceMat;
-    }
-  };
-
-  template <typename Scalar, typename LO, typename GO>
-  class ForwardImportMatrix {
-  private:
-    using crs_type = CrsMatrix<Scalar, LO, GO>;
-
-  public:
-    void operator()(RCP<crs_type> sourceMat, RCP<crs_type> targetMat) const {
-      Import<LO, GO> importer(sourceMat->getMap(), targetMat->getMap(), getImportParameterList());
-      targetMat->doImport(*sourceMat, importer, INSERT);
-      targetMat->fillComplete();
-    }
-  };
-
-  TEUCHOS_UNIT_TEST_TEMPLATE_3_DECL( CrsMatrixTransfer, asyncImport_diagonal, LO, GO, Scalar )
-  {
-    CrsMatrixDiagonalTransferFixture<Scalar, LO, GO> fixture(out, success);
-    if (fixture.shouldSkipTest()) return;
-
-    fixture.setupMatrices(0);
-    fixture.performTransfer(ForwardImportMatrix<Scalar, LO, GO>());
-    fixture.checkResults(ReferenceImportMatrix<Scalar, LO, GO>());
-  }
-
-
-  template <typename Scalar, typename LO, typename GO>
-  class CrsMatrixLowerTriangularTransferFixture {
-  private:
-    using map_type = Map<LO, GO>;
-    using crs_type = CrsMatrix<Scalar, LO, GO>;
-
-  public:
-    CrsMatrixLowerTriangularTransferFixture(Teuchos::FancyOStream& o, bool& s)
-      : out(o),
-        success(s),
-        comm(getDefaultComm()),
-        numProcs(comm->getSize()),
-        myRank(comm->getRank())
-    { }
-
-    ~CrsMatrixLowerTriangularTransferFixture() { }
-
-    bool shouldSkipTest() {
-      return hasOddNumProcs();
-    }
-
-    void setupMatrices() {
-      const GO indexBase = 0;
-      const Tpetra::global_size_t INVALID = Teuchos::OrdinalTraits<Tpetra::global_size_t>::invalid();
-
-      const size_t sourceNumLocalElements = (myRank%2 == 0) ? 3 : 5;
-      const size_t targetNumLocalElements = 4;
-
-      sourceMap = rcp(new map_type(INVALID, sourceNumLocalElements, indexBase, comm));
-      sourceMat = rcp(new crs_type(sourceMap, 24, StaticProfile, getCrsMatrixParameterList()));
-
-      Array<GO> cols(1);
-      Array<Scalar>  vals(1);
-      for (GO row = sourceMap->getMinGlobalIndex(); row <= sourceMap->getMaxGlobalIndex(); row++) {
-        if (row > 0) {
-          cols.resize(row);
-          vals.resize(row);
-          for (GO col=0; col<row; col++) {
-            cols[col] = as<GO>(col);
-            vals[col] = as<Scalar>(col);
-          }
-          sourceMat->insertGlobalValues(row, cols, vals);
-        }
-      }
-      sourceMat->fillComplete();
-
-      targetMap = rcp(new map_type(INVALID, targetNumLocalElements, indexBase, comm));
-      targetMat = rcp(new crs_type(targetMap, 24, StaticProfile, getCrsMatrixParameterList()));
-    }
-
-    template <typename TransferMethod>
-    void performTransfer(const TransferMethod& transfer) {
-      transfer(sourceMat, targetMat);
-    }
-
-    void checkResults() {
-      using lids_type = typename crs_type::local_inds_host_view_type;
-      using vals_type = typename crs_type::values_host_view_type;
-
-      const RCP<const map_type> colMap = targetMat->getColMap();
-
-      for (GO globalRow=targetMap->getMinGlobalIndex(); globalRow<=targetMap->getMaxGlobalIndex(); ++globalRow) {
-        LO localRow = targetMap->getLocalElement(globalRow);
-        lids_type rowIndices;
-        vals_type rowValues;
-
-        targetMat->getLocalRowView(localRow, rowIndices, rowValues);
-        TEST_EQUALITY(rowIndices.extent(0), (size_t) globalRow);
-        TEST_EQUALITY(rowValues.extent(0), (size_t) globalRow);
-
-        Array<GO> indices(rowIndices.size());
-        Array<Scalar> values(rowValues.size());
-
-        for (decltype (rowIndices.size()) j=0; j<rowIndices.size(); ++j) {
-          indices[j] = colMap->getGlobalElement(rowIndices[j]);
-          values[j] = rowValues[j];
-        }
-        Tpetra::sort2(indices.begin(), indices.end(), values.begin());
-
-        for (decltype (rowIndices.size()) j=0; j<rowIndices.size(); ++j) {
-          TEST_EQUALITY(indices[j], as<GO>(j));
-          TEST_EQUALITY(values[j], as<Scalar>(j));
-        }
-      }
-    }
-
-  private:
-    bool hasOddNumProcs() {
-      if (numProcs%2 != 0) {
-        out << "This test is only meaningful if running with an even number of MPI processes." << endl;
-        return true;
-      }
-      else return false;
-    }
-
-    Teuchos::FancyOStream& out;
-    bool& success;
-
-    RCP<const Comm<int>> comm;
-    const int numProcs;
-    const int myRank;
-
-    RCP<const map_type> sourceMap;
-    RCP<const map_type> targetMap;
-
-    RCP<crs_type> sourceMat;
-    RCP<crs_type> targetMat;
-  };
-
-  TEUCHOS_UNIT_TEST_TEMPLATE_3_DECL( CrsMatrixTransfer, asyncImport_lowerTriangular, LO, GO, Scalar )
-  {
-    CrsMatrixLowerTriangularTransferFixture<Scalar, LO, GO> fixture(out, success);
-    if (fixture.shouldSkipTest()) return;
-
-    fixture.setupMatrices();
-    fixture.performTransfer(ForwardImportMatrix<Scalar, LO, GO>());
-    fixture.checkResults();
-  }
-
-
 template<class Graph>
 bool graphs_are_same(const RCP<Graph>& G1, const RCP<const Graph>& G2)
 {
@@ -3469,9 +3038,6 @@ TEUCHOS_UNIT_TEST_TEMPLATE_2_DECL( Import_Util,GetTwoTransferOwnershipVector, LO
 #define UNIT_TEST_GROUP_SC_LO_GO( SC, LO, GO )                   \
   TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( CrsMatrixImportExport, doImport, LO, GO, SC ) \
   TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( MultiVectorImport, doImport, LO, GO, SC ) \
-  TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( MultiVectorTransfer, asyncImport, LO, GO, SC ) \
-  TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( CrsMatrixTransfer, asyncImport_diagonal, LO, GO, SC ) \
-  TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( CrsMatrixTransfer, asyncImport_lowerTriangular, LO, GO, SC ) \
   TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( FusedImportExport, doImport, LO, GO, SC ) \
   TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( Import_Util, UnpackAndCombineWithOwningPIDs, LO, GO, SC ) \
   TEUCHOS_UNIT_TEST_TEMPLATE_3_INSTANT( FusedImportExport, MueLuStyle, LO, GO, SC )
